Refact: Separate menuItems component into a function
